+++ examples/resources/morpheus_credential/resource_api_key.tf
@@ -1,0 +1,7 @@
+resource "morpheus_credential" "tf_example_credential_api_key" {
+  name        = "tf_example_credential_api_key"
+  description = "terraform credential example for api key"
+  enabled     = true
+  type        = "api-key"
+  api_key     = "FIEFMIQNQ"
+}